pyproject.toml
[project]
name = "model-quantize-and-dequantize"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pillow>=11.3.0",
    "seaborn>=0.13.2",
    "torch>=2.7.1",
    "tornado>=6.5.1",
    "transformers>=4.54.0",
]
